fix: sign-extend hash_bytes tail bytes, and restore the id author string
nnue_hash_bytes zero-extended its trailing bytes. Upstream's hash_bytes takes
`const char*` and does `u64(end[i])` (misc.cpp:481); plain char is SIGNED on
x86-64, so a tail byte >= 0x80 fills the high bits with ones. ccfish therefore
produced a different digest for any input whose length is not a multiple of 8
and whose tail has the high bit set. The eval-file identity hash runs over the
net's description string, which can carry such bytes.

The `id author` handshake line read "(see AUTHORS)" where upstream says
"(see AUTHORS file)" -- a transcription slip on a line GUIs log, live in the
shipped binary.

Bit-exactness verified to hold, and the net still loads and hashes correctly:

  signature  2466447, unchanged
  eval       NNUE evaluation using nn-0ee0657fb25e.nnue (106MiB, ...)
  perft      unchanged
  test       1464474 checks, 0 failures, ASan+UBSan
  ./build.sh parity -> all gates passed

Found by a hostile line-by-line audit of the shell decomposition (2447 lines).
That audit verified all 19 UCI options entry-by-entry against upstream's
registration order, defaults, and bounds, and confirmed the live search facade's
hardcoded option values match -- exonerating the option table. Its remaining
findings are landmines in unwired files: engine.c's search_clear does not clear
histories and its EvalFile callback does not load the net, both with comments
claiming those subsystems are unported when they are live elsewhere.

+++ src/engine/eval/nnue/nnue_hash.c
@@ -40,7 +40,12 @@ uint64_t nnue_hash_bytes(const uint8_t *data, size_t len) {
         size_t i = len & 7;
         while (i > 0) {
             i -= 1;
-            k = (k << 8) | (uint64_t) data[tail + i];
+            // SIGN-extend. Upstream's hash_bytes takes `const char*` and does
+            // `u64(end[i])` (misc.cpp:481); plain char is SIGNED on x86-64, so a
+            // tail byte >= 0x80 fills the high bits with ones. Zero-extending
+            // here would silently produce a different digest for any input whose
+            // length is not a multiple of 8 and whose tail has the high bit set.
+            k = (k << 8) | (uint64_t) (int64_t) (int8_t) data[tail + i];
         }
         h ^= k;
         h *= m;
